Move BdvUIPanel and related classes to package bdv.util.uipanel
