+++ src/main/java/bdv/util/BdvSource.java
@@ -48,8 +48,8 @@ public abstract class BdvSource implements Bdv
 	{
 		return numTimepoints;
 	}
-	
-	protected abstract String getName() ;
 
-	protected abstract String getTypeAsString() ;
+	public abstract String getName(); // TODO
+
+ 	public abstract String getTypeAsString(); // TODO
 }
