+++ src/main/resources/bdv/util/uipanel/notVisible.png
