+++ src/main/resources/bdv/util/uipanel/notVisible_small.png
