+++ src/main/resources/bdv/util/uipanel/upbutton.png
