+++ src/main/resources/bdv/util/uipanel/visible.png
