+++ src/main/resources/bdv/util/uipanel/visible_small.png
